- SIEP Link fix - Email upgarde - Routes rename - MAJ Parrainnage - Make it responsive :-)

+++ frontend/src/services/requests/faction.service.ts
@@ -1,19 +1,19 @@
 import api from '../api';
 
 export const getAllFactions = async () => {
-  const res = await api.get("faction/factions");
+  const res = await api.get("faction/admin/factions");
 
   return res.data.data;
 };
 
 export const createFaction = async (factionName : string) => {
-  const res = await api.post("faction/createfaction", {factionName});
+  const res = await api.post("faction/admin/createfaction", {factionName});
 
   return res.data;
 };
 
 export const deleteFaction = async (factionId : number) => {
-  const res = await api.delete("faction/deletefaction", {params : {factionId}});
+  const res = await api.delete("faction/admin/deletefaction", {params : {factionId}});
 
   return res.data;
 };
